Refine dtype & utils.py

+++ ocnn/octree/octree.py
@@ -274,7 +274,7 @@ class Octree:
       children[0] = 0
 
     # update octree
-    self.children[depth] = children
+    self.children[depth] = children.int()
     self.nnum_nempty[depth] = nnum_nempty
 
   def octree_grow(self, depth: int, update_neigh: bool = True):
